requirements.txt
numpy
opencv-python
scikit-image
matplotlib
python-dotenv